requirements.txt
openai
python-dotenv
langchain_community
langchain_text_splitters
pypdf
pymongo
langchain